# apps/api/docs/swagger.yaml
definitions:
  github_com_projectx_api_internal_platform_httpx.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
info:
  contact:
    name: Project-X support
  description: |-
    REST API for Project-X — the AI-run marketing + studio operations platform.
    Auth is a JWT stored in an HTTP-only cookie (set by POST /api/v1/auth/login).
    When calling protected endpoints from a tool like Swagger UI or curl, log in first and reuse the returned cookie.
  license:
    name: Proprietary
  title: Project-X API
  version: "1.0"
paths:
...
  /api/v1/me/studios/{id}/billing/upgrade:
    post:
      consumes:
      - application/json
      description: Looks up the requested tier's price from the platform plans setting
        (or hardcoded defaults) and creates a Stripe Checkout session for the studio
        to pay for/upgrade to that plan. Uses subscription mode for recurring tiers,
        or one-time payment mode for the "Trial Pass" tier. The session is tagged
        with plan_tier/studio_id/is_upgrade metadata so the Stripe webhook can finalize
        the upgrade (and cancel any superseded subscription) once payment completes.
        Returns the Checkout session URL to redirect the studio to.
      parameters:
      - description: Studio ID
        in: path
        name: id
        required: true
        type: string
      - description: Target plan tier name
        in: body
        name: body
        required: true
        schema:
          properties:
            tier:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: global studio, invalid studio ID, invalid request body, or
            invalid subscription tier
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "404":
          description: studio not found
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: platform Stripe account not configured, or Stripe error creating
            the checkout session
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Upgrade or purchase a studio's subscription plan
      tags:
      - Billing
...
  /api/v1/me/studios/global/plans:
    put:
      consumes:
      - application/json
      description: Replaces the "platform_plans" platform setting with the given list
        of plans. Intended for the platform owner to edit pricing shown to prospective
        studios; note the route is not itself role-restricted beyond requiring authentication.
      parameters:
      - description: Full list of plan objects to save
        in: body
        name: body
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update platform subscription plans
      tags:
      - Platform
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: JWT session cookie set by POST /api/v1/auth/login (cookie name configurable
      via COOKIE_NAME).
    in: cookie
    name: px_session
    type: apiKey
swagger: "2.0"